requirements.txt
python-docx
streamlit
langchain
faiss-cpu
sentence-transformers
pypdf2
unstructured
openai
requests
beautifulsoup4
langchain-community